# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: CFA API
  version: 1.0.0
servers:
  - url: /
...
paths:
  /user/register:
    post:
      tags:
        - User
      operationId: register_post
      requestBody:
        $ref: '#/components/requestBodies/register_user_dto'
        content:
          application/json:
            schema:
              properties:
                login:
                  type: string
                username:
                  type: string
                name:
                  type: string
                password:
                  type: string
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/login_response_200'
          description: Успешный ответ
        "400":
          description: Некорректный запрос
      summary: Зарегистрировать нового пользователя
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/login:
    post:
      tags:
        - User
      operationId: login_post
      requestBody:
        $ref: '#/components/requestBodies/login_user_dto'
        content:
          application/json:
            schema:
              properties:
                login:
                  type: string
                password:
                  type: string
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/login_response_200'
          description: Успешный ответ
        "401":
          description: Неверные учетные данные
      summary: Войти как пользователь
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/profile:
    get:
      tags:
        - User
      operationId: profile_get
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Получить информацию о своём профиле
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/{user-id}:
    get:
      tags:
        - User
      operationId: user_user_id_get
      parameters:
        - explode: false
          in: path
          name: user-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
          description: Успешный ответ
        "404":
          description: Пользователь не найден
      summary: Получить информацию о профиле другого пользователя
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/deposit:
    post:
      tags:
        - User
      operationId: user_deposit_post
      requestBody:
        $ref: '#/components/requestBodies/deposit_value_dto'
        content:
          application/json:
            schema:
              properties:
                count:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Пополнить баланс
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/withdraw:
    post:
      tags:
        - User
      operationId: user_withdraw_post
      requestBody:
        $ref: '#/components/requestBodies/withdraw_value_dto'
        content:
          application/json:
            schema:
              properties:
                count:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Вывести деньги с баланса
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/cfa/{user-id}:
    get:
      tags:
        - User
      operationId: user_cfa_user_id_get
      parameters:
        - explode: false
          in: path
          name: user-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/user_cfa_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Пользователь не найден
      summary: Получить список CFA для пользователя
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/offer/{user-id}:
    get:
      tags:
        - User
      operationId: user_offer_user_id_get
      parameters:
        - explode: false
          in: path
          name: user-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/offer_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Пользователь не найден
      summary: Получить список всех своих предложений
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/desire/{user-id}:
    get:
      tags:
        - User
      operationId: user_desire_user_id_get
      parameters:
        - explode: false
          in: path
          name: user-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/desire_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Пользователь не найден
      summary: Получить список всех своих предложений
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/deal/in:
    get:
      tags:
        - User
      operationId: user_deal_in_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/deal_dto'
                type: array
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
        "404":
          description: Пользователь не найден
      security:
        - BearerAuth: [ ]
      summary: Получить список всех обменов, направленных пользователю
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /user/deal/out:
    get:
      tags:
        - User
      operationId: user_deal_out_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/deal_dto'
                type: array
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
        "404":
          description: Пользователь не найден
      security:
        - BearerAuth: [ ]
      summary: Получить список всех обменов, созданных пользователем
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /cfa-image/create:
    post:
      tags:
        - CfaImage
      operationId: cfa_image_create_post
      requestBody:
        $ref: '#/components/requestBodies/create_cfa_image_dto'
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                count:
                  type: integer
                description:
                  type: string
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfa_image_id_dto'
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Создать новый CFA Image (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
...
  /cfa-image/price/{cfa-image-id}:
    get:
      tags:
        - CfaImage
      operationId: cfa_image_price_cfa_image_id_get
      parameters:
        - explode: false
          in: path
          name: cfa-image-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfa_price_dto'
          description: Успешный ответ
        "404":
          description: Изображение CFA не найдено
      summary: Получить цену изображения CFA на основе рыночных предложений
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /cfa-image/pred-price/{cfa-image-id}:
    get:
      tags:
        - CfaImage
      operationId: cfa_image_predict_price
      parameters:
        - explode: false
          in: path
          name: cfa-image-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  type: number
                type: array
          description: Успешный ответ
        "404":
          description: Изображение CFA не найдено
      summary: Получить цену изображения CFA на основе рыночных предложений
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /cfa/{cfa-token}:
    get:
      tags:
        - Cfa
      operationId: cfa_cfa_token_get
      parameters:
        - explode: false
          in: path
          name: cfa-token
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cfa_dto'
          description: Успешный ответ
        "404":
          description: CFA не найден
      summary: Получить информацию о CFA по его токену
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /cfa/list/{cfa-image-id}:
    get:
      tags:
        - Cfa
      operationId: cfa_list_cfa_image_id_get
      parameters:
        - explode: false
          in: path
          name: cfa-image-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/cfa_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Изображение CFA не найдено
      summary: Получить список CFAs для изображения CFA
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /cfa/history/{cfa-token}:
    get:
      tags:
        - Cfa
      operationId: cfa_history_cfa_token_get
      parameters:
        - explode: false
          in: path
          name: cfa-token
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/trade_dto'
                type: array
          description: Успешный ответ
        "404":
          description: CFA не найден
      summary: Получить историю сделок для CFA по его токену
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /offer/create:
    post:
      tags:
        - Offer
      operationId: offer_create_post
      requestBody:
        $ref: '#/components/requestBodies/create_offer_dto'
        content:
          application/json:
            schema:
              properties:
                cfa_image_id:
                  type: integer
                count:
                  type: integer
                price:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Создать новое предложение (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /offer/cancel/{offer-id}:
    post:
      tags:
        - Offer
      operationId: offer_cancel_offer_id_post
      parameters:
        - explode: false
          in: path
          name: offer-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Удалить предложение (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /offer/buy/{offer-id}:
    post:
      tags:
        - Offer
      operationId: offer_buy_offer_id_post
      parameters:
        - explode: false
          in: path
          name: offer-id
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        $ref: '#/components/requestBodies/accept_offer_dto'
        content:
          application/json:
            schema:
              properties:
                count:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
        "404":
          description: Предложение не найдено
        "409":
          description: Недостаточно товара для покупки
      security:
        - BearerAuth: [ ]
      summary: Купить предложение (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /offer/list/{cfa-image-id}:
    get:
      tags:
        - Offer
      operationId: offer_list_cfa_image_id_get
      parameters:
        - explode: false
          in: path
          name: cfa-image-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/offer_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Изображение CFA не найдено
      summary: Получить список предложений для изображения CFA
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /desire/create:
    post:
      tags:
        - Desire
      operationId: desire_create_post
      requestBody:
        $ref: '#/components/requestBodies/create_desire_dto'
        content:
          application/json:
            schema:
              properties:
                cfa_image_id:
                  type: integer
                count:
                  type: integer
                price:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Создать новое предложение (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /desire/cancel/{desire-id}:
    post:
      tags:
        - Desire
      operationId: desire_cancel_desire_id_post
      parameters:
        - explode: false
          in: path
          name: desire-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Удалить предложение (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /desire/sell/{desire-id}:
    post:
      tags:
        - Desire
      operationId: desire_sell_desire_id_post
      parameters:
        - explode: false
          in: path
          name: desire-id
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        $ref: '#/components/requestBodies/accept_desire_dto'
        content:
          application/json:
            schema:
              properties:
                count:
                  type: integer
              type: object
        required: true
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
        "404":
          description: Предложение не найдено
        "409":
          description: Недостаточно товара для покупки
      security:
        - BearerAuth: [ ]
      summary: Продать cfa (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
  /desire/list/{cfa-image-id}:
    get:
      tags:
        - Desire
      operationId: desire_list_cfa_image_id_get
      parameters:
        - explode: false
          in: path
          name: cfa-image-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/desire_dto'
                type: array
          description: Успешный ответ
        "404":
          description: Изображение CFA не найдено
      summary: Получить список желаний для изображения CFA
      x-openapi-router-controller: openapi_server.controllers.default_controller
...
  /deal/cancel/{deal-id}:
    post:
      tags:
        - Deal
      operationId: deal_cancel_deal_id_post
      parameters:
        - explode: false
          in: path
          name: deal-id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          description: Успешный ответ
        "401":
          description: Неверный Bearer-токен
      security:
        - BearerAuth: [ ]
      summary: Удалить обмен (Требуется Bearer-токен)
      x-openapi-router-controller: openapi_server.controllers.default_controller
...
components:
  requestBodies:
    deposit_value_dto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/deposit_value_dto'
      required: true
    withdraw_value_dto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/withdraw_value_dto'
      required: true
    login_user_dto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/login_user_dto'
      required: true
    register_user_dto:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/register_user_dto'
      required: true
...
  schemas:
    User:
      properties:
        id:
          type: integer
        login:
          type: string
        username:
          type: string
        name:
          type: string
        balance:
          type: number
          format: float
      type: object
    PublicUser:
      properties:
        id:
          type: integer
        login:
          type: string
        username:
          type: string
        name:
          type: string
      type: object
...
    user_cfa_dto:
      properties:
        cfa_image:
          $ref: '#/components/schemas/PublicUser'
        count:
          type: integer
        tokens:
          type: array
          items:
            type: string
...
    login_response_200:
      properties:
        user:
          $ref: '#/components/schemas/User'
        bearer_token:
          type: string
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
      x-bearerInfoFunc: openapi_server.controllers.security_controller_.info_from_BearerAuth